refactor: main.tsの再生/描画/ABC変換ロジックを分離して責務を整理 ## 概要 `main.ts` に集中していた大きな処理を分離し、UI層と処理層の責務を明確化しました。 あわせて、ABC→MusicXML変換ロジックを `abc-io.ts` に集約しました。
## 変更内容

### 1. 再生フローを `playback-flow.ts` に分離
- 新規追加: `src/ts/playback-flow.ts`
- 移動対象:
  - `createBasicWaveSynthEngine`
  - `startPlayback`
  - `startMeasurePlayback`
  - `stopPlayback` 相当処理
- `main.ts` は `PlaybackFlowOptions` 経由で UI 更新・状態更新コールバックを渡す構成に変更

### 2. Verovio描画フローを `preview-flow.ts` に分離
- 新規追加: `src/ts/preview-flow.ts`
- 移動対象:
  - `renderScorePreview` の描画本体
  - `renderMeasureEditorPreview` の描画本体
- `main.ts` は表示先DOM・マップ更新・ハイライト処理を渡す薄いラッパー化

### 3. ABC変換ロジックを `abc-io.ts` に集約
- `src/ts/abc-io.ts` に移動/追加:
  - `convertAbcToMusicXml`
  - `buildMusicXmlFromAbcParsed`
  - `xmlEscape`
  - `normalizeTypeForMusicXml`
  - `clefXmlFromAbcClef`（`export` 化）
- `main.ts` 側の重複定義（`AbcParsed*` 型や変換関数群）を削除

### 4. `main.ts` のimport/呼び出しを整理
- `playback-flow.ts` / `preview-flow.ts` / `abc-io.ts` の関数を利用
- 旧実装の直接依存を削除して可読性を改善

## 変更ファイル
- `src/ts/playback-flow.ts` (new)
- `src/ts/preview-flow.ts` (new)
- `src/ts/abc-io.ts`
- `src/ts/main.ts`
- `src/js/main.js`
- `mikuscore.html`

## 効果
- `main.ts` の肥大化を抑制し、保守性を向上
- 再生・描画・ABC変換を個別モジュールで追える構成に改善
- 今後の機能追加時に影響範囲を限定しやすくなる

## 備考
- 本PRは主にリファクタ
